Move supporting_methods to init; Move messages there as well
